# api-docs .yaml
openapi: 3.0.3
info:
  title: 无人机道路病害检测系统 API
  description: |
    这是《无人机道路病害检测》大创项目的核心后端接口文档。
    涵盖了无人机设备管理、任务调度、AI 数据接入与地图瓦片渲染。

    **【实时通信 WebSocket 指南】**
    * **连接地址**: `ws://<服务器IP>:8080/ws/telemetry`
    * **推送事件**: 
      1. `type: "telemetry"` - 无人机高频移动坐标
      2. `type: "new_defect"` - 突发病害报警 (含坐标和图片)
      3. `type: "task_status_update"` - 任务完成与归位通知
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: 本地开发环境

tags:
  - name: 巡检任务管理
    description: 巡检任务的创建、状态流转与综合查询
  - name: 地图可视化
    description: 前端高德地图渲染专用的瓦片加载与懒加载详情接口
  - name: 数据采集与上报
    description: 供无人机/Python模拟器调用的高频数据与图片上传入口
  - name: 无人机设备管理
    description: 无人机台账的增删改查

paths:
  # ==========================================
  # 巡检任务管理
  # ==========================================
  /api/v1/tasks:
    get:
      tags:
        - 巡检任务管理
      summary: 获取任务列表
      operationId: getTaskList
      parameters:
        - name: taskName
          in: query
          description: 任务名称(模糊查询)
          schema:
            type: string
        - name: status
          in: query
          description: 状态 (0-未开始, 1-执行中, 2-已完成)
          schema:
            type: integer
        - name: droneId
          in: query
          description: 无人机ID
          schema:
            type: integer
        - name: defectCountMin
          in: query
          description: 最小病害数量
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListInspectionTask'
    post:
      tags:
        - 巡检任务管理
      summary: 创建新任务
      operationId: taskCreate
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionTask'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInspectionTask'

  /api/v1/tasks/{id}:
    get:
      tags:
        - 巡检任务管理
      summary: 获取单条任务详情
      operationId: getTask
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInspectionTask'
    put:
      tags:
        - 巡检任务管理
      summary: 修改任务信息
      operationId: taskUpdate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InspectionTask'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInspectionTask'
    delete:
      tags:
        - 巡检任务管理
      summary: 删除任务
      operationId: taskDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultInspectionTask'

  /api/v1/tasks/{id}/start:
    put:
      tags:
        - 巡检任务管理
      summary: 开始任务
      description: 前端调用，将任务和无人机状态设为 1。
      operationId: taskStart
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'

  /api/v1/tasks/{droneId}/finish:
    put:
      tags:
        - 巡检任务管理
      summary: 结束任务
      description: 模拟器飞完后调用，将任务设为 2 (已完成)，无人机设为 0 (空闲)，并广播 WS 通知。
      operationId: taskFinish
      parameters:
        - name: droneId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultString'

  # ==========================================
  # 地图可视化
  # ==========================================
  /api/v1/map/tile:
    get:
      tags:
        - 地图可视化
      summary: 获取地图瓦片内的病害图片
      description: 前端拖拽地图时，传入当前屏幕覆盖的瓦片 XYZ 坐标。后端通过 Redis 锁并发处理。
      operationId: getDefectsInViewport
      parameters:
        - name: z
          in: query
          required: true
          description: 缩放级别 Zoom
          schema:
            type: integer
        - name: x
          in: query
          required: true
          schema:
            type: integer
        - name: "y"
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功返回该瓦片内的图片点
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListInspectionImage'

  /api/v1/map/{imageId}/details:
    get:
      tags:
        - 地图可视化
      summary: 懒加载某张图片的病害明细
      description: 在大屏点击某张图片时触发，返回该图中识别出的具体病害列表。
      operationId: getDetail
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListDefectDetail'

  # ==========================================
  # 数据采集与上报
  # ==========================================
  /api/v1/drones/upload:
    post:
      tags:
        - 数据采集与上报
      summary: 图片抓拍与 AI 识别
      operationId: uploadImage
      parameters:
        - name: taskId
          in: query
          required: true
          schema:
            type: integer
        - name: droneId
          in: query
          required: true
          schema:
            type: integer
        - name: lng
          in: query
          required: true
          schema:
            type: number
            format: double
        - name: lat
          in: query
          required: true
          schema:
            type: number
            format: double
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: 抓拍的原图文件
      responses:
        "200":
          description: 成功处理
          content:
            application/json:
              schema:
                type: string

  /api/v1/drones/telemetry:
    post:
      tags:
        - 数据采集与上报
      summary: 高频遥测坐标上报
      operationId: receiveTelemetry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                taskId:
                  type: integer
                droneId:
                  type: integer
                lng:
                  type: number
                  format: double
                lat:
                  type: number
                  format: double
                heading:
                  type: number
                  format: double
        required: true
      responses:
        "200":
          description: 广播成功
          content:
            application/json:
              schema:
                type: string

  # ==========================================
  # 无人机设备管理
  # ==========================================
  /api/v1/devices:
    get:
      tags:
        - 无人机设备管理
      summary: 获取无人机台账列表
      operationId: getDroneList
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultListDroneDevice'
    post:
      tags:
        - 无人机设备管理
      summary: 新增无人机
      operationId: droneUpload
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DroneDevice'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDroneDevice'

  /api/v1/devices/{id}:
    get:
      tags:
        - 无人机设备管理
      summary: 获取单台无人机信息
      operationId: getDrone
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDroneDevice'
    put:
      tags:
        - 无人机设备管理
      summary: 更新无人机信息
      operationId: droneUpdate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DroneDevice'
        required: true
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDroneDevice'
    delete:
      tags:
        - 无人机设备管理
      summary: 报废/删除无人机
      operationId: droneDelete
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResultDroneDevice'

components:
  schemas:
    InspectionTask:
      type: object
      properties:
        id:
          type: integer
        taskName:
          type: string
          example: "南区干道自动巡检"
        droneId:
          type: integer
        routePoints:
          type: string
          description: "前端保存的航点坐标，JSON字符串格式"
          example: '[{"lng": 116.397, "lat": 39.909}, {"lng": 116.402, "lat": 39.905}]'
        status:
          type: integer
          description: "0-未开始, 1-执行中, 2-已完成"
          example: 0
        defectCount:
          type: integer
          description: "病害统计数量"
        createTime:
          type: string
          format: date-time
        updateTime:
          type: string
          format: date-time

    DroneDevice:
      type: object
      properties:
        id:
          type: integer
        droneName:
          type: string
          example: "大疆 M300-RTK-01"
        status:
          type: integer
          description: "0-空闲, 1-任务中"
          example: 0
        lastLng:
          type: number
          format: double
        lastLat:
          type: number
          format: double
        createTime:
          type: string
          format: date-time

    InspectionImage:
      type: object
      properties:
        id:
          type: integer
        taskId:
          type: integer
        droneId:
          type: integer
        originalImageUrl:
          type: string
          example: "/origin/2026-03-05/xyz.jpg"
        resultImageUrl:
          type: string
          example: "/result/2026-03-05/xyz.jpg"
        matchedLng:
          type: number
          format: double
        matchedLat:
          type: number
          format: double
        isDefect:
          type: integer
          description: "1-有病害, 0-无病害"
        defectCount:
          type: integer
        captureTime:
          type: string
          format: date-time

    DefectDetail:
      type: object
      properties:
        id:
          type: integer
        imageId:
          type: integer
        defectType:
          type: string
          example: "crack"
        confidence:
          type: number
          format: double
          example: 0.95
        bbox:
          type: string
          description: "病害矩形框的坐标数组"
          example: "[10, 20, 100, 200]"

    # ====== 统一返回包装类 ======
    ResultString:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
          example: "操作成功"
        data:
          type: string

    ResultInspectionTask:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
        data:
          $ref: '#/components/schemas/InspectionTask'

    ResultListInspectionTask:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/InspectionTask'

    ResultDroneDevice:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
        data:
          $ref: '#/components/schemas/DroneDevice'

    ResultListDroneDevice:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/DroneDevice'

    ResultListInspectionImage:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/InspectionImage'

    ResultListDefectDetail:
      type: object
      properties:
        code:
          type: integer
          example: 200
        msg:
          type: string
        data:
          type: array
          items:
            $ref: '#/components/schemas/DefectDetail'